requirements.txt
addict==2.4.0
appdirs==1.4.4
asttokens==3.0.0
autocommand==2.2.2
backports.tarfile==1.2.0
blessed==1.27.0
causal_conv1d==1.5.0.post8
certifi==2022.12.7
charset-normalizer==2.1.1
cityscapesScripts==2.2.4
coloredlogs==15.0.1
colormaps==0.4.2
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
fast_pytorch_kmeans==0.2.2
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gpustat==1.1.1
hf-xet==1.1.10
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.4
importlib_metadata==8.0.0
inflect==7.3.1
iopath==0.1.10
ipython==8.37.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
Jinja2==3.1.4
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mmengine==0.10.7
mmsegmentation==0.11.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.3
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==13.590.48
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==24.2
parso==0.8.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
portalocker==3.2.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.9.18
requests==2.28.1
rich==14.2.0
safetensors==0.6.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
timm==0.4.12
tokenizers==0.22.1
tomli==2.0.1
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
triton==2.2.0
typeguard==4.3.0
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==1.26.13
wcwidth==0.2.14
yacs==0.1.8
yapf==0.40.1
zipp==3.19.2
